requirements.txt
Django>=1.7
djangorestframework
markdown  # For DRF browsable API.
pathlib2; python_version < "3"
